+++ agixt/Agent.py
@@ -1100,132 +1100,3 @@ class Agent:
 - **THE ASSISTANT CANNOT EXECUTE A COMMAND THAT IS NOT ON THE LIST OF EXAMPLES!**"""
             return agent_commands
         return ""
-
-    async def write_text_to_memory(
-        self,
-        user_input: str,
-        text: str,
-        conversation_id: str = None,
-        external_source: str = "user input",
-    ):
-        session = get_session()
-        try:
-            # Remove old content for external sources
-            if external_source.startswith(("file", "http://", "https://")):
-                session.query(Memory).filter_by(
-                    agent_id=self.agent_id,
-                    conversation_id=conversation_id,
-                    external_source=external_source,
-                ).delete()
-
-            # Split into chunks and embed
-            chunks = await self.chunk_content(text=text, chunk_size=self.chunk_size)
-            if self.summarize_content:
-                chunks = [await self.summarize_text(chunk) for chunk in chunks]
-
-            for chunk in chunks:
-                embedding = self.embeddings(chunk)
-                memory = Memory(
-                    agent_id=self.agent_id,
-                    conversation_id=conversation_id,
-                    embedding=embedding,
-                    text=chunk,
-                    external_source=external_source,
-                    description=user_input,
-                    additional_metadata=chunk,
-                )
-                session.add(memory)
-
-            session.commit()
-            return True
-        except Exception as e:
-            session.rollback()
-            logging.error(f"Error writing to memory: {e}")
-            return False
-        finally:
-            session.close()
-
-    async def get_memories(
-        self,
-        user_input: str,
-        limit: int = 5,
-        min_relevance_score: float = 0.0,
-        conversation_id: str = None,
-    ):
-        session = get_session()
-        DATABASE_TYPE = getenv("DATABASE_TYPE")
-        try:
-            query_embedding = self.embeddings(user_input)
-
-            if DATABASE_TYPE == "sqlite":
-                # SQLite VSS query
-                results = session.execute(
-                    """
-                    SELECT m.*, distance
-                    FROM memory m
-                    WHERE m.agent_id = :agent_id
-                    AND (m.conversation_id = :conversation_id OR m.conversation_id IS NULL)
-                    AND vss_memories MATCH :embedding
-                    ORDER BY distance DESC
-                    LIMIT :limit
-                """,
-                    {
-                        "agent_id": self.agent_id,
-                        "conversation_id": conversation_id,
-                        "embedding": str(query_embedding.tolist()),
-                        "limit": limit,
-                    },
-                )
-            else:
-                # PostgreSQL vector query
-                results = session.execute(
-                    """
-                    SELECT m.*, 
-                        1 - (m.embedding <=> :embedding::vector) as similarity
-                    FROM memory m
-                    WHERE m.agent_id = :agent_id
-                    AND (m.conversation_id = :conversation_id OR m.conversation_id IS NULL)
-                    ORDER BY m.embedding <=> :embedding::vector
-                    LIMIT :limit
-                """,
-                    {
-                        "agent_id": self.agent_id,
-                        "conversation_id": conversation_id,
-                        "embedding": query_embedding.tolist(),
-                        "limit": limit,
-                    },
-                )
-
-            memories = []
-            for row in results:
-                score = row.distance if DATABASE_TYPE == "sqlite" else row.similarity
-                if score >= min_relevance_score:
-                    memories.append(
-                        {
-                            "text": row.text,
-                            "external_source_name": row.external_source,
-                            "description": row.description,
-                            "additional_metadata": row.additional_metadata,
-                            "relevance_score": score,
-                        }
-                    )
-
-            return memories
-        finally:
-            session.close()
-
-    async def wipe_memory(self, conversation_id: str = None):
-        session = get_session()
-        try:
-            query = session.query(Memory).filter_by(agent_id=self.agent_id)
-            if conversation_id:
-                query = query.filter_by(conversation_id=conversation_id)
-            query.delete()
-            session.commit()
-            return True
-        except Exception as e:
-            session.rollback()
-            logging.error(f"Error wiping memory: {e}")
-            return False
-        finally:
-            session.close()
